+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.4.20
+- show `Alan Show` button when `ALAN_APP_URL` is set
+- show `Alan Deploy` button when `ALAN_CONTAINER_NAME` is set
+
 ## 0.4.19
 - support Alan platform versions >= 2021.8